Promote canonical schema IR to shared crate per Issue 572 tdl R3
- Add shared `map_schema_semantic` crate for source-neutral schema IR, literals, diagnostics, and symbol indexing
- Bridge map-schema JSON literals into the shared literal abstraction
- Document semantic boundaries around source formats, inverse metadata, variants, and derived indexes
- Refresh generated schema/TDL outputs for the R3 canonical IR flow
