+++ schema-src/MAP Schema Types-map-core-schema-operator-types.tdl
@@ -19,6 +19,7 @@ abstract holon OperatorType {
 relationship AffordsOperator {
   source ValueType
   target OperatorType.HolonType
+  inverse AffordedBy
   cardinality 0..32767
   header {
     description: "Links a ValueType to the OperatorTypes available for that value type, supporting validation and query construction."
@@ -32,7 +33,6 @@ relationship AffordsOperator {
 inverse relationship AffordedBy {
   source OperatorType.HolonType
   target ValueType
-  inverse AffordsOperator
   cardinality 1..32767
   header {
     description: "Inverse of AffordsOperator, relationship from OperatorType to ValueType."
@@ -53,31 +53,15 @@ enum OperatorCategory {
   }
   variants {
     variant Equality {
-      properties {
-        type_name: "Equality"
-        display_name: "Equality"
+      header {
         description: "Operators that test whether two values of the same type are equal or unequal. Applicable to all scalar value types."
-        instance_type_kind: "TypeKind.EnumVariant"
-        is_abstract_type: false
-      }
-      relationships {
-        ComponentOf -> "MAP Core Schema-v0.0.7"
-        Extends -> "MapEnumVariantValueType"
-        VariantOf -> "OperatorCategory"
+        display_name: "Equality"
       }
     }
     variant Ordering {
-      properties {
-        type_name: "Ordering"
-        display_name: "Ordering"
+      header {
         description: "Operators that compare the relative order of two values of the same type. Applicable only to value types with a defined total order (e.g., integer, string)."
-        instance_type_kind: "TypeKind.EnumVariant"
-        is_abstract_type: false
-      }
-      relationships {
-        ComponentOf -> "MAP Core Schema-v0.0.7"
-        Extends -> "MapEnumVariantValueType"
-        VariantOf -> "OperatorCategory"
+        display_name: "Ordering"
       }
     }
   }
