+++ shared_crates/map_schema_semantic/src/schema_ir.rs
@@ -1,0 +1,546 @@
+//! Canonical Holon IR for MAP schema tooling.
+//!
+//! This module owns the descriptor-aware semantic graph shared by the current
+//! decompiler, compiler, symbol index, and round-trip comparison tests. It is
+//! intentionally syntax-neutral: JSON import parsing and TDL parsing both lower
+//! into these structures before derived lookup indexes are built.
+
+use crate::{
+    literal_value::LiteralObject,
+    schema_index::{SymbolId, SymbolIndex},
+};
+use std::fmt;
+use std::{collections::BTreeMap, path::PathBuf};
+
+/// Semantic descriptor categories represented in the Canonical Holon IR.
+///
+/// These are descriptor roles, not parser productions. Source adapters should normalize any
+/// format-specific spelling into this vocabulary before symbol indexing or validation runs.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub enum DescriptorKind {
+    Schema,
+    TypeDescriptor,
+    HolonType,
+    PropertyType,
+    RelationshipType,
+    ValueType,
+    Enum,
+    EnumVariant,
+}
+
+impl fmt::Display for DescriptorKind {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Schema => write!(f, "Schema"),
+            Self::TypeDescriptor => write!(f, "TypeDescriptor"),
+            Self::HolonType => write!(f, "HolonType"),
+            Self::PropertyType => write!(f, "PropertyType"),
+            Self::RelationshipType => write!(f, "RelationshipType"),
+            Self::ValueType => write!(f, "ValueType"),
+            Self::Enum => write!(f, "Enum"),
+            Self::EnumVariant => write!(f, "EnumVariant"),
+        }
+    }
+}
+
+/// Distinguishes declared relationship descriptors from inverse relationship descriptors.
+///
+/// The flavor comes from authored schema intent. It is kept separately from relationship-pair
+/// metadata so tools can preserve whether a descriptor is the declared side or inverse side even
+/// when both sides reference one another.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+pub enum RelationshipFlavor {
+    /// Authored relationship descriptor that owns the relationship-pair declaration.
+    Declared,
+    /// Authored inverse descriptor paired with a declared relationship.
+    Inverse,
+}
+
+/// Source family for origin metadata.
+///
+/// Origins are diagnostic/provenance data only. They must not affect semantic equivalence.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SourceKind {
+    /// Descriptor was lowered from canonical JSON import data.
+    JsonImport,
+    /// Descriptor was lowered from TDL source.
+    TdlSource,
+    /// Descriptor was generated by tooling rather than directly authored.
+    Generated,
+}
+
+/// Optional source location attached to schemas, descriptors, symbols, and diagnostics.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Origin {
+    pub source_kind: SourceKind,
+    pub file_path: Option<PathBuf>,
+    pub line: Option<u32>,
+    pub column: Option<u32>,
+}
+
+impl Origin {
+    /// Creates origin metadata when only the source family is known.
+    pub fn new(source_kind: SourceKind) -> Self {
+        Self { source_kind, file_path: None, line: None, column: None }
+    }
+
+    /// Creates file-level origin metadata for a JSON import source.
+    pub fn json_file(path: impl Into<PathBuf>) -> Self {
+        Self {
+            source_kind: SourceKind::JsonImport,
+            file_path: Some(path.into()),
+            line: None,
+            column: None,
+        }
+    }
+
+    /// Creates source-location metadata for a TDL source file.
+    pub fn tdl_file(path: impl Into<PathBuf>, line: Option<u32>, column: Option<u32>) -> Self {
+        Self { source_kind: SourceKind::TdlSource, file_path: Some(path.into()), line, column }
+    }
+}
+
+/// Role a semantic reference plays on its owning descriptor.
+///
+/// A role describes why a target is referenced, which lets validation enforce role-specific target
+/// kinds and lets projections choose the correct loader relationship. It is intentionally broader
+/// than any one source syntax.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub enum ReferenceRole {
+    /// Schema membership relationship.
+    ComponentOf,
+    /// Type inheritance relationship.
+    Extends,
+    /// Key rule used by a descriptor.
+    KeyRule,
+    /// Source endpoint of a relationship type.
+    SourceType,
+    /// Target endpoint of a relationship type.
+    TargetType,
+    /// Inverse-side pointer back to its declared relationship.
+    InverseOf,
+    /// Declared-side pointer to its inverse relationship.
+    HasInverse,
+    /// Value type used by a property type.
+    ValueType,
+    /// Enum-to-variant membership declared on the enum descriptor.
+    Variants,
+    /// Variant-to-enum membership retained when a source format provides that shape.
+    VariantOf,
+    /// Holon/type instance property membership.
+    InstanceProperty,
+    /// Holon/type instance relationship membership.
+    InstanceRelationship,
+    /// Schema dependency relationship.
+    DependsOn,
+}
+
+impl fmt::Display for ReferenceRole {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::ComponentOf => write!(f, "ComponentOf"),
+            Self::Extends => write!(f, "Extends"),
+            Self::KeyRule => write!(f, "UsesKeyRule"),
+            Self::SourceType => write!(f, "SourceType"),
+            Self::TargetType => write!(f, "TargetType"),
+            Self::InverseOf => write!(f, "InverseOf"),
+            Self::HasInverse => write!(f, "HasInverse"),
+            Self::ValueType => write!(f, "ValueType"),
+            Self::Variants => write!(f, "Variants"),
+            Self::VariantOf => write!(f, "VariantOf"),
+            Self::InstanceProperty => write!(f, "InstanceProperties"),
+            Self::InstanceRelationship => write!(f, "InstanceRelationships"),
+            Self::DependsOn => write!(f, "DependsOn"),
+        }
+    }
+}
+
+/// Reference text as authored/imported, optionally resolved to a symbol.
+///
+/// `target` preserves the source-facing reference text for diagnostics and round-trip comparison.
+/// `resolved` is derived by [`SymbolIndex`] and should be treated as build-local, not persisted.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SemanticReference {
+    pub role: ReferenceRole,
+    pub target: String,
+    pub resolved: Option<SymbolId>,
+}
+
+impl SemanticReference {
+    /// Creates a reference that has not yet been resolved against a symbol index.
+    pub fn unresolved(role: ReferenceRole, target: impl Into<String>) -> Self {
+        Self { role, target: target.into(), resolved: None }
+    }
+
+    /// Attaches a build-local symbol ID after semantic indexing.
+    pub fn resolve(&mut self, symbol_id: SymbolId) {
+        self.resolved = Some(symbol_id);
+    }
+
+    /// Returns a stable comparison key that ignores derived symbol IDs.
+    pub fn comparable_key(&self) -> (&ReferenceRole, &str) {
+        (&self.role, &self.target)
+    }
+}
+
+/// Human-facing descriptor header fields preserved for later emission.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct DescriptorHeader {
+    pub description: Option<String>,
+    pub display_name: Option<String>,
+    pub display_name_plural: Option<String>,
+    pub type_name_plural: Option<String>,
+}
+
+/// Literal relationship entries preserved from holon-shaped source inputs.
+///
+/// These are relationship-like values that have not been promoted into first-class semantic slots.
+/// Keeping them separate prevents accidental validation as canonical descriptor references.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct LiteralRelationship {
+    pub name: String,
+    pub targets: Vec<String>,
+}
+
+/// Semantic representation of a schema declaration.
+///
+/// A schema is a descriptor container plus source-preserved metadata. Canonical references such as
+/// dependencies are modeled explicitly; additional literal properties/relationships are retained
+/// for loss-aware projection without becoming validation-driving semantics.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Schema {
+    pub name: String,
+    pub key: String,
+    pub origin: Origin,
+    pub dependencies: Vec<SemanticReference>,
+    pub literal_properties: LiteralObject,
+    pub literal_relationships: Vec<LiteralRelationship>,
+    pub header: Option<DescriptorHeader>,
+    pub allows_additional_properties: bool,
+    pub allows_additional_relationships: bool,
+}
+
+/// Semantic representation of a non-schema descriptor.
+///
+/// This structure is intentionally descriptor-aware rather than a generic bag of relationships.
+/// Common semantic references get dedicated slots so validators, decompilers, and loader
+/// projections can agree on meaning without reinterpreting literal relationship names.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct TypeDescriptor {
+    pub key: String,
+    pub name: String,
+    pub kind: DescriptorKind,
+    pub owning_schema: String,
+    pub origin: Origin,
+    pub header: Option<DescriptorHeader>,
+    pub is_abstract: bool,
+    pub extends: Option<SemanticReference>,
+    pub component_of: Option<SemanticReference>,
+    pub key_rule: Option<SemanticReference>,
+    pub value_type: Option<SemanticReference>,
+    pub source_type: Option<SemanticReference>,
+    pub target_type: Option<SemanticReference>,
+    /// Inverse-side relationship-pair metadata pointing to the declared relationship.
+    pub inverse_of: Option<SemanticReference>,
+    /// Declared-side relationship-pair metadata pointing to the inverse relationship.
+    pub has_inverse: Option<SemanticReference>,
+    /// Enum-owned links to variant descriptors.
+    pub variants: Vec<SemanticReference>,
+    /// Variant-owned link to its enum when preserved from a source shape.
+    pub variant_of: Option<SemanticReference>,
+    pub literal_properties: LiteralObject,
+    pub instance_properties: Vec<SemanticReference>,
+    pub instance_relationships: Vec<SemanticReference>,
+    pub literal_relationships: Vec<LiteralRelationship>,
+    pub relationship_flavor: Option<RelationshipFlavor>,
+    pub is_definitional: bool,
+    pub min_cardinality: Option<i64>,
+    pub max_cardinality: Option<i64>,
+    pub deletion_semantic: Option<String>,
+    pub is_ordered: bool,
+    pub allows_duplicates: bool,
+    pub allows_additional_properties: bool,
+    pub allows_additional_relationships: bool,
+}
+
+impl TypeDescriptor {
+    /// Creates a descriptor with no optional semantic fields populated.
+    pub fn new(
+        key: impl Into<String>,
+        name: impl Into<String>,
+        kind: DescriptorKind,
+        owning_schema: impl Into<String>,
+        origin: Origin,
+    ) -> Self {
+        Self {
+            key: key.into(),
+            name: name.into(),
+            kind,
+            owning_schema: owning_schema.into(),
+            origin,
+            header: None,
+            is_abstract: false,
+            extends: None,
+            component_of: None,
+            key_rule: None,
+            value_type: None,
+            source_type: None,
+            target_type: None,
+            inverse_of: None,
+            has_inverse: None,
+            variants: Vec::new(),
+            variant_of: None,
+            literal_properties: LiteralObject::new(),
+            instance_properties: Vec::new(),
+            instance_relationships: Vec::new(),
+            literal_relationships: Vec::new(),
+            relationship_flavor: None,
+            is_definitional: false,
+            min_cardinality: None,
+            max_cardinality: None,
+            deletion_semantic: None,
+            is_ordered: false,
+            allows_duplicates: false,
+            allows_additional_properties: false,
+            allows_additional_relationships: false,
+        }
+    }
+
+    /// Returns all semantic references carried by this descriptor.
+    ///
+    /// Literal relationships are intentionally excluded because they have not been normalized into
+    /// canonical semantic roles.
+    pub fn references(&self) -> Vec<&SemanticReference> {
+        let mut refs = Vec::new();
+        refs.extend(self.extends.iter());
+        refs.extend(self.component_of.iter());
+        refs.extend(self.key_rule.iter());
+        refs.extend(self.value_type.iter());
+        refs.extend(self.source_type.iter());
+        refs.extend(self.target_type.iter());
+        refs.extend(self.inverse_of.iter());
+        refs.extend(self.has_inverse.iter());
+        refs.extend(self.variants.iter());
+        refs.extend(self.variant_of.iter());
+        refs.extend(self.instance_properties.iter());
+        refs.extend(self.instance_relationships.iter());
+        refs
+    }
+
+    /// Returns all semantic references carried by this descriptor for resolution.
+    pub fn references_mut(&mut self) -> Vec<&mut SemanticReference> {
+        let mut refs = Vec::new();
+        refs.extend(self.extends.iter_mut());
+        refs.extend(self.component_of.iter_mut());
+        refs.extend(self.key_rule.iter_mut());
+        refs.extend(self.value_type.iter_mut());
+        refs.extend(self.source_type.iter_mut());
+        refs.extend(self.target_type.iter_mut());
+        refs.extend(self.inverse_of.iter_mut());
+        refs.extend(self.has_inverse.iter_mut());
+        refs.extend(self.variants.iter_mut());
+        refs.extend(self.variant_of.iter_mut());
+        refs.extend(self.instance_properties.iter_mut());
+        refs.extend(self.instance_relationships.iter_mut());
+        refs
+    }
+}
+
+/// A source-normalized collection of schemas and descriptors.
+///
+/// This is the canonical input to derived services. Callers should mutate this model, then rebuild
+/// indexes rather than treating indexes as authoritative state.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SemanticModel {
+    pub schemas: Vec<Schema>,
+    pub descriptors: Vec<TypeDescriptor>,
+}
+
+impl SemanticModel {
+    /// Creates an empty semantic model.
+    pub fn new() -> Self {
+        Self { schemas: Vec::new(), descriptors: Vec::new() }
+    }
+
+    /// Adds a schema declaration to the model in source order.
+    pub fn push_schema(&mut self, schema: Schema) {
+        self.schemas.push(schema);
+    }
+
+    /// Adds a descriptor declaration to the model in source order.
+    pub fn push_descriptor(&mut self, descriptor: TypeDescriptor) {
+        self.descriptors.push(descriptor);
+    }
+
+    /// Resolves known references in-place against an already-derived symbol table.
+    pub fn resolve_references(&mut self, symbols: &SymbolIndex) {
+        for schema in &mut self.schemas {
+            for reference in &mut schema.dependencies {
+                if let Some(symbol) = symbols.lookup_reference_target(&reference.target) {
+                    reference.resolve(symbol.id);
+                }
+            }
+        }
+
+        for descriptor in &mut self.descriptors {
+            for reference in descriptor.references_mut() {
+                if let Some(symbol) = symbols.lookup_reference_target(&reference.target) {
+                    reference.resolve(symbol.id);
+                }
+            }
+        }
+    }
+
+    /// Produces a stable, origin-insensitive signature for round-trip comparisons.
+    pub fn comparable_signature(&self) -> ComparableSemanticModel {
+        let mut schemas = self
+            .schemas
+            .iter()
+            .map(|schema| ComparableSchema {
+                name: schema.name.clone(),
+                key: schema.key.clone(),
+                dependencies: comparable_refs(&schema.dependencies),
+                allows_additional_properties: schema.allows_additional_properties,
+                allows_additional_relationships: schema.allows_additional_relationships,
+            })
+            .collect::<Vec<_>>();
+        schemas.sort();
+
+        let mut descriptors = self
+            .descriptors
+            .iter()
+            .map(|descriptor| ComparableDescriptor {
+                key: descriptor.key.clone(),
+                name: descriptor.name.clone(),
+                kind: descriptor.kind,
+                owning_schema: descriptor.owning_schema.clone(),
+                is_abstract: descriptor.is_abstract,
+                references: comparable_descriptor_refs(descriptor),
+                relationship_flavor: descriptor.relationship_flavor,
+                is_definitional: descriptor.is_definitional,
+                min_cardinality: descriptor.min_cardinality,
+                max_cardinality: descriptor.max_cardinality,
+                deletion_semantic: descriptor.deletion_semantic.clone(),
+                is_ordered: descriptor.is_ordered,
+                allows_duplicates: descriptor.allows_duplicates,
+                allows_additional_properties: descriptor.allows_additional_properties,
+                allows_additional_relationships: descriptor.allows_additional_relationships,
+            })
+            .collect::<Vec<_>>();
+        descriptors.sort();
+
+        ComparableSemanticModel { schemas, descriptors }
+    }
+}
+
+impl Default for SemanticModel {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+/// Origin-insensitive semantic model representation for equivalence checks.
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct ComparableSemanticModel {
+    pub schemas: Vec<ComparableSchema>,
+    pub descriptors: Vec<ComparableDescriptor>,
+}
+
+/// Origin-insensitive schema representation for equivalence checks.
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct ComparableSchema {
+    pub name: String,
+    pub key: String,
+    pub dependencies: Vec<(ReferenceRole, String)>,
+    pub allows_additional_properties: bool,
+    pub allows_additional_relationships: bool,
+}
+
+/// Origin-insensitive descriptor representation for equivalence checks.
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct ComparableDescriptor {
+    pub key: String,
+    pub name: String,
+    pub kind: DescriptorKind,
+    pub owning_schema: String,
+    pub is_abstract: bool,
+    pub references: Vec<(ReferenceRole, String)>,
+    pub relationship_flavor: Option<RelationshipFlavor>,
+    pub is_definitional: bool,
+    pub min_cardinality: Option<i64>,
+    pub max_cardinality: Option<i64>,
+    pub deletion_semantic: Option<String>,
+    pub is_ordered: bool,
+    pub allows_duplicates: bool,
+    pub allows_additional_properties: bool,
+    pub allows_additional_relationships: bool,
+}
+
+fn comparable_descriptor_refs(descriptor: &TypeDescriptor) -> Vec<(ReferenceRole, String)> {
+    let mut refs = descriptor
+        .references()
+        .into_iter()
+        .map(|reference| (reference.role, reference.target.clone()))
+        .collect::<Vec<_>>();
+    refs.sort();
+    refs
+}
+
+fn comparable_refs(references: &[SemanticReference]) -> Vec<(ReferenceRole, String)> {
+    let mut refs = references
+        .iter()
+        .map(|reference| (reference.role, reference.target.clone()))
+        .collect::<Vec<_>>();
+    refs.sort();
+    refs
+}
+
+/// Minimal adapter input for future TDL AST-to-semantic tests and loaders.
+///
+/// This keeps early parser tests from constructing full descriptors by hand while still forcing
+/// them through the same role-based reference placement used by production adapters.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct TdlDescriptorInput {
+    pub key: String,
+    pub name: String,
+    pub kind: DescriptorKind,
+    pub owning_schema: String,
+    pub origin: Origin,
+    pub references: BTreeMap<ReferenceRole, Vec<String>>,
+}
+
+impl TdlDescriptorInput {
+    /// Converts adapter input into a fully slotted descriptor.
+    pub fn into_descriptor(self) -> TypeDescriptor {
+        let mut descriptor =
+            TypeDescriptor::new(self.key, self.name, self.kind, self.owning_schema, self.origin);
+        for (role, targets) in self.references {
+            for target in targets {
+                push_reference(&mut descriptor, SemanticReference::unresolved(role, target));
+            }
+        }
+        descriptor
+    }
+}
+
+/// Adds a reference to the matching semantic slot on a descriptor.
+///
+/// Single-valued roles overwrite the previous value; multi-valued roles append. `DependsOn` is a
+/// schema-level role and is ignored for descriptors.
+pub fn push_reference(descriptor: &mut TypeDescriptor, reference: SemanticReference) {
+    match reference.role {
+        ReferenceRole::ComponentOf => descriptor.component_of = Some(reference),
+        ReferenceRole::Extends => descriptor.extends = Some(reference),
+        ReferenceRole::KeyRule => descriptor.key_rule = Some(reference),
+        ReferenceRole::SourceType => descriptor.source_type = Some(reference),
+        ReferenceRole::TargetType => descriptor.target_type = Some(reference),
+        ReferenceRole::InverseOf => descriptor.inverse_of = Some(reference),
+        ReferenceRole::HasInverse => descriptor.has_inverse = Some(reference),
+        ReferenceRole::Variants => descriptor.variants.push(reference),
+        ReferenceRole::ValueType => descriptor.value_type = Some(reference),
+        ReferenceRole::VariantOf => descriptor.variant_of = Some(reference),
+        ReferenceRole::InstanceProperty => descriptor.instance_properties.push(reference),
+        ReferenceRole::InstanceRelationship => descriptor.instance_relationships.push(reference),
+        ReferenceRole::DependsOn => {}
+    }
+}
