+++ tools/map-schema/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1"
 clap = { version = "4.5.16", features = ["derive"] }
+map_schema_semantic = { path = "../../shared_crates/map_schema_semantic" }
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["preserve_order"] }
 time = { version = "=0.3.36", features = ["formatting"] }
